Move album page to be SSR
