+++ components/homePage.tsx
@@ -1,0 +1,5 @@
+export default function HomePage() {
+    return (
+        <div>HomePage</div>
+    )
+}
