+++ next.config.ts
@@ -3,8 +3,7 @@ import createMDX from "@next/mdx";
 
 const nextConfig: NextConfig = {
   transpilePackages: ["next-mdx-remote"],
-  pageExtensions: ["js", "jsx", "md", "mdx", "ts", "tsx"],
-  devIndicators: false
+  pageExtensions: ["js", "jsx", "md", "mdx", "ts", "tsx"]
 };
 
 const withMDX = createMDX({
